pyproject.toml
[tool.poetry]
name = "wikipedia-tests"
version = "0.1.0"
description = ""
authors = ["lrayne <legaterayne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
appium-python-client = "^4.0.1"
pytest = "^8.3.1"
allure-pytest = "^2.13.5"
selene = {version = "^2.0.0rc9", allow-prereleases = true}
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.pytest.ini_options]
addopts = [
    '--clean-alluredir',
    '--alluredir=allure-results'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
